+++ packages/editor/src/mometa/runtime/provider.tsx
@@ -3,16 +3,18 @@ import { ConfigProvider } from 'antd'
 import zhCN from 'antd/lib/locale/zh_CN'
 import React from 'react'
 import { DndProvider } from 'react-dnd'
+import getWindow from 'get-window'
 
 import { getSharedFromMain } from '../utils/get-from-main'
 const { _sharedMap, _sharedUpdateMap, ddManager } = getSharedFromMain()
 
-export function Provider({ children }) {
+export function Provider({ children, dom }) {
   React.useLayoutEffect(() => {
     const bk: any = ddManager.getBackend()
-    bk.addEventListeners(window)
+    const win = getWindow(dom)
+    bk.addEventListeners(win)
     return () => {
-      bk.removeEventListeners(window)
+      bk.removeEventListeners(win)
     }
   }, [ddManager])