+++ packages/react-refresh-webpack-plugin/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.18](https://github.com/pmmmwh/react-refresh-webpack-plugin/compare/v0.0.17...v0.0.18) (2022-01-07)
+
+**Note:** Version bump only for package @mometa/react-refresh-webpack-plugin
+
+## [0.0.17](https://github.com/pmmmwh/react-refresh-webpack-plugin/compare/v0.0.16...v0.0.17) (2022-01-07)
+
+**Note:** Version bump only for package @mometa/react-refresh-webpack-plugin
+
 ## [0.0.16](https://github.com/pmmmwh/react-refresh-webpack-plugin/compare/v0.0.15...v0.0.16) (2022-01-04)
 
 **Note:** Version bump only for package @mometa/react-refresh-webpack-plugin